backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
kiteconnect==5.0.1
python-dotenv==1.0.1
gunicorn==23.0.0
